+++ python/MaterialXTest/tests_to_html.py
@@ -14,18 +14,32 @@ except Exception:
 
 def createDiff(image1Path, image2Path, imageDiffPath):
     try:
+        if os.path.exists(imageDiffPath):
+            os.remove(imageDiffPath)
+        
+        if not os.path.exists(image1Path):
+            print ("Image diff input missing: " + image1Path)
+            return
+        
+        if not os.path.exists(image2Path):
+            print ("Image diff input missing: " + image2Path)
+            return
+
         image1 = Image.open(image1Path).convert('RGB')
         image2 = Image.open(image2Path).convert('RGB')
         diff = ImageChops.difference(image1, image2)
         diff.save(imageDiffPath)
     except Exception:
+        if os.path.exists(imageDiffPath):
+            os.remove(imageDiffPath)
         print ("Failed to create image diff between: " + image1Path + ", " + image2Path)
 
 def main(args=None):
 
     parser = argparse.ArgumentParser()
-    parser.add_argument('-i', '--inputdir', dest='inputdir', action='store', help='Input directory', default=".")
+    parser.add_argument('-i1', '--inputdir1', dest='inputdir1', action='store', help='Input directory', default=".")
     parser.add_argument('-i2', '--inputdir2', dest='inputdir2', action='store', help='Second input directory', default=".")
+    parser.add_argument('-i3', '--inputdir3', dest='inputdir3', action='store', help='Third input directory', default=".")
     parser.add_argument('-o', '--outputfile', dest='outputfile', action='store', help='Output file name', default="tests.html")
     parser.add_argument('-d', '--diff', dest='CREATE_DIFF', action='store_true', help='Perform image diff', default=False)
     parser.add_argument('-t', '--timestamp', dest='ENABLE_TIMESTAMPS', action='store_true', help='Write image timestamps', default=False)
@@ -33,8 +47,9 @@ def main(args=None):
     parser.add_argument('-ht', '--imageheight', type=int, dest='imageheight', action='store', help='Set image display height', default=256)
     parser.add_argument('-cp', '--cellpadding', type=int, dest='cellpadding', action='store', help='Set table cell padding', default=0)
     parser.add_argument('-tb', '--tableborder', type=int, dest='tableborder', action='store', help='Table border width. 0 means no border', default=3)
-    parser.add_argument('-sl', '--sourcelang', dest='sourcelang', action='store', help='Source language. Default is source', default="glsl")
-    parser.add_argument('-dl', '--destlang', dest='destlang', action='store', help='Destination language. Default is dest', default="osl")
+    parser.add_argument('-l1', '--lang1', dest='lang1', action='store', help='First target language for comparison. Default is glsl', default="glsl")
+    parser.add_argument('-l2', '--lang2', dest='lang2', action='store', help='Second target language for comparison. Default is osl', default="osl")
+    parser.add_argument('-l3', '--lang3', dest='lang3', action='store', help='Third target language for comparison. Default is empty', default="")
 
     args = parser.parse_args(args)
 
@@ -52,84 +67,132 @@ def main(args=None):
     fh.write("}")
     fh.write("</style>")
     fh.write("<body>\n")
-    dir1 = os.getcwd() if args.inputdir == "." else args.inputdir
+    dir1 = os.getcwd() if args.inputdir1 == "." else args.inputdir1
     dir2 = os.getcwd() if args.inputdir2 == "." else args.inputdir2
-    fh.write("<h3>" + args.sourcelang + " (in: " + dir1 + ") vs "+ args.destlang + " (in: " + dir2 + ")</h3>\n")
+    dir3 = os.getcwd() if args.inputdir3 == "." else args.inputdir3
+    
+    useThirdLang = True if args.lang3 and (args.inputdir1 != args.inputdir3 or args.lang1 != args.lang3) else False
+    
+    if useThirdLang:
+        fh.write("<h3>" + args.lang1 + " (in: " + dir1 + ") vs "+ args.lang2 + " (in: " + dir2 + ") vs "+ args.lang3 + " (in: " + dir3 + ")</h3>\n")
+    else:
+        fh.write("<h3>" + args.lang1 + " (in: " + dir1 + ") vs "+ args.lang2 + " (in: " + dir2 + ")</h3>\n")
 
     if not DIFF_ENABLED and args.CREATE_DIFF:
         print("--diff argument ignored. Diff utility not installed.")
 
     if not args.inputdir2:
-        args.inputdir2 = args.inputdir
+        args.inputdir2 = args.inputdir1
+    if useThirdLang and not args.inputdir3:
+        args.inputdir3 = args.inputdir1
 
     # Get all source files
-    sourceFiles = []
-    sourcePaths = []
-    for subdir, _, files in os.walk(args.inputdir):
+    langFiles1 = []
+    langPaths1 = []
+    for subdir, _, files in os.walk(args.inputdir1):
         for curFile in files:
-            if curFile.endswith(args.sourcelang + ".png"):
-                sourceFiles.append(curFile)
-                sourcePaths.append(subdir) 
+            if curFile.endswith(args.lang1 + ".png"):
+                langFiles1.append(curFile)
+                langPaths1.append(subdir) 
 
     # Get all destination files, matching source files
-    destFiles = []
-    destPaths = []
-    postFix = args.sourcelang + ".png"
-    for sourceFile, sourcePath in zip(sourceFiles, sourcePaths):
+    langFiles2 = []
+    langPaths2 = []
+    langFiles3 = []
+    langPaths3 = []
+    postFix = args.lang1 + ".png"
+    for file1, path1 in zip(langFiles1, langPaths1):
         # Allow for just one language to be shown if source and dest are the same.
         # Otherwise add in equivalent name with dest language replacement if
         # pointing to the same directory 
-        if args.inputdir != args.inputdir2 or args.sourcelang != args.destlang:
-            destFile = sourceFile[:-len(postFix)] + args.destlang + ".png"
-            destPath = os.path.join(args.inputdir2, sourcePath)
+        if args.inputdir1 != args.inputdir2 or args.lang1 != args.lang2:
+            file2 = file1[:-len(postFix)] + args.lang2 + ".png"
+            path2 = os.path.join(args.inputdir2, path1)
         else:
-            destFile = ""
-            destPath = None
-        destFiles.append(destFile)
-        destPaths.append(destPath)
+            file2 = ""
+            path2 = None
+        langFiles2.append(file2)
+        langPaths2.append(path2)
 
-    if sourceFiles:
+        if useThirdLang:
+            file3 = file1[:-len(postFix)] + args.lang3 + ".png"
+            path3 = os.path.join(args.inputdir2, path1)
+        else:
+            file3 = ""
+            path3 = None
+        langFiles3.append(file3)
+        langPaths3.append(path3)
+
+    if langFiles1:
         curPath = ""
-        for sourceFile, destFile, sourcePath, destPath in zip(sourceFiles, destFiles, sourcePaths, destPaths):
+        for file1, file2, file3, path1, path2, path3 in zip(langFiles1, langFiles2, langFiles3, langPaths1, langPaths2, langPaths3):
 
-            fullSourcePath = os.path.join(sourcePath, sourceFile) if sourceFile else None
-            fullDestPath = os.path.join(destPath, destFile) if destFile else None
+            fullPath1 = os.path.join(path1, file1) if file1 else None
+            fullPath2 = os.path.join(path2, file2) if file2 else None
+            fullPath3 = os.path.join(path3, file3) if file3 else None
+            diffPath1 = None
+            diffPath2 = None
+            diffPath3 = None
 
-            if curPath != sourcePath:
+            if curPath != path1:
                 if curPath != "":
                     fh.write("</table>\n")
-                fh.write("<p>" + os.path.normpath(sourcePath) + ":</p>\n")
+                fh.write("<p>" + os.path.normpath(path1) + ":</p>\n")
                 fh.write("<table>\n")
-                curPath = sourcePath
+                curPath = path1
 
-            if sourceFile and destFile and DIFF_ENABLED and args.CREATE_DIFF:
-                diffPath = fullSourcePath[0:-8] + "diff.png"
-                createDiff(fullSourcePath, fullDestPath, diffPath)
+            if file1 and file2 and DIFF_ENABLED and args.CREATE_DIFF:
+                diffPath1 = fullPath1[0:-8] + "_" + args.lang1 + "_vs_" + args.lang2 + "_diff.png"
+                createDiff(fullPath1, fullPath2, diffPath1)
+
+            if useThirdLang and file1 and file3 and DIFF_ENABLED and args.CREATE_DIFF:
+                diffPath2 = fullPath1[0:-8] + "_" + args.lang1 + "_vs_" + args.lang3 + "_diff.png"
+                createDiff(fullPath1, fullPath3, diffPath2)
+                diffPath3 = fullPath1[0:-8] + "_" + args.lang2 + "_vs_" + args.lang3 + "_diff.png"
+                createDiff(fullPath2, fullPath3, diffPath3)
+
+            if os.path.isabs(args.outputfile):
+                fileUri = 'file:///'
             else:
-                diffPath = None
+                fileUri =''
 
             fh.write("<tr>\n")
-            if fullSourcePath:
-                fh.write("<td class='td_image'><img src='" + fullSourcePath + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
-            if fullDestPath:
-                fh.write("<td class='td_image'><img src='" + fullDestPath + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
-            if diffPath:
-                fh.write("<td class='td_image'><img src='" + diffPath + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if fullPath1:
+                fh.write("<td class='td_image'><img src='" + fileUri + fullPath1 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if fullPath2:
+                fh.write("<td class='td_image'><img src='" + fileUri + fullPath2 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if fullPath3:
+                fh.write("<td class='td_image'><img src='" + fileUri + fullPath3 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if diffPath1:
+                fh.write("<td class='td_image'><img src='" + fileUri + diffPath1 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if diffPath2:
+                fh.write("<td class='td_image'><img src='" + fileUri + diffPath2 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
+            if diffPath3:
+                fh.write("<td class='td_image'><img src='" + fileUri + diffPath3 + "' height='" + str(args.imageheight) + "' width='" + str(args.imagewidth) + "' loading='lazy' style='background-color:black;'/></td>\n")
             fh.write("</tr>\n")
 
             fh.write("<tr>\n")
-            if fullSourcePath:
-                fh.write("<td align='center'>" + sourceFile)
-            if args.ENABLE_TIMESTAMPS and os.path.isfile(fullSourcePath):
-                fh.write("<br>(" + str(datetime.datetime.fromtimestamp(os.path.getmtime(fullSourcePath))) + ")")
-            fh.write("</td>\n")
-            if fullDestPath:
-                fh.write("<td align='center'>" + destFile)
-            if args.ENABLE_TIMESTAMPS and os.path.isfile(fullDestPath):
-                fh.write("<br>(" + str(datetime.datetime.fromtimestamp(os.path.getmtime(fullDestPath))) + ")")
-            fh.write("</td>\n")
-            if diffPath:
-                fh.write("<td align='center'>Difference</td>\n")
+            if fullPath1:
+                fh.write("<td align='center'>" + file1)
+                if args.ENABLE_TIMESTAMPS and os.path.isfile(fullPath1):
+                    fh.write("<br>(" + str(datetime.datetime.fromtimestamp(os.path.getmtime(fullPath1))) + ")")
+                fh.write("</td>\n")
+            if fullPath2:
+                fh.write("<td align='center'>" + file2)
+                if args.ENABLE_TIMESTAMPS and os.path.isfile(fullPath2):
+                    fh.write("<br>(" + str(datetime.datetime.fromtimestamp(os.path.getmtime(fullPath2))) + ")")
+                fh.write("</td>\n")
+            if fullPath3:
+                fh.write("<td align='center'>" + file3)
+                if args.ENABLE_TIMESTAMPS and os.path.isfile(fullPath3):
+                    fh.write("<br>(" + str(datetime.datetime.fromtimestamp(os.path.getmtime(fullPath3))) + ")")
+                fh.write("</td>\n")
+            if diffPath1:
+                fh.write("<td align='center'>Difference " + args.lang1 + " vs. " + args.lang2 + " </td>\n")
+            if diffPath2:
+                fh.write("<td align='center'>Difference " + args.lang1 + " vs. " + args.lang3 + " </td>\n")
+            if diffPath3:
+                fh.write("<td align='center'>Difference " + args.lang2 + " vs. " + args.lang3 + " </td>\n")
             fh.write("</tr>\n")
 
     fh.write("</table>\n")